第三版

+++ hw2/0716061_hw2c.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 
 int main(void) {
-    long long day, month, year;
+    int day, month, year;
     printf("Please enter a data (format: mm/dd/yyyy): ");
-    scanf("%lld/%lld/%lld", &month, &day, &year);
-    printf("your input data is %04lld%02lld%02lld.\n", year, month, day);
+    scanf("%d/%d/%d", &month, &day, &year);
+    printf("your input data is %04d%02d%02d.\n", year, month, day);
     return 0;
 }
